Fix cross-cypress-test import in node cves
